requirements.txt
python-telegram-bot~=21.5
pytz~=2023.3.post1

colorlog~=6.7.0
aiohttp~=3.9.1
pyzbar~=0.1.9
numpy~=2.1.1
Pillow~=10.4.0
matplotlib~=3.9.2
beautifulsoup4~=4.12.3
pycryptodome~=3.20.0
qrcode~=7.4.2
